将rtp时间戳改为32位

+++ src/Rtsp/Rtsp.cpp
@@ -558,8 +558,8 @@ uint32_t RtpPacket::getStamp() const {
     return ntohl(getHeader()->stamp);
 }
 
-uint64_t RtpPacket::getStampMS(bool ntp) const {
-    return ntp ? ntp_stamp : getStamp() * uint64_t(1000) / sample_rate;
+uint32_t RtpPacket::getStampMS(bool ntp) const {
+    return ntp ? ntp_stamp : getStamp() * (uint32_t)1000 / sample_rate;
 }
 
 uint32_t RtpPacket::getSSRC() const {
